refactor(describegpt): extract DuckDB / SQL cluster into submodule
Move the DuckDB discovery and SQL helpers into src/cmd/describegpt/duckdb_sql.rs:
escape_sql_string, extract_sql_sample, should_use_duckdb, get_duckdb_path,
run_duckdb_query, score_sql_query, build_score_refinement_prompt, and
handle_sql_error. Update session.rs to import extract_sql_sample from its
new location.

Also tighten the escape_sql_string docs to say explicitly that it is only
safe for `'…'` string literals, not identifiers — this was a secondary
finding from the Pass-1 review and the clearer docstring is worth keeping
even though no separate ident-escape helper is needed yet (no current caller
splices user-controlled identifiers).

Main file drops from 4,622 to 4,399 lines — cumulative reduction across
this branch: 1,177 lines (21%).
Clippy clean; all 64 integration + 7 unit tests pass.
